refactor(replay): 修复批量验证消费侧摩擦点的边界清理
依据 docs/review/batch_validation_consumer_friction.md 的复核发现，处理
初版"最小变动修复"留下的 4 处边界问题（边界 2 SAR schema wire 兼容性
因项目未上线暂不处理）：

边界 1（§1 分叉映射字符串匹配）：ReplayTraceOutputCallback 返回类型
从 bool 改为结构化枚举 ReplayTraceOutputStatus
（kHandledByModule/kDivergence/kOtherFailure），删除 IsModuleOutputDivergenceError
字符串匹配；四模块 OnCycleOutput 显式返回三态。统一两套分叉路径的
stop_on_first_divergence 语义与 compared_output_count 计数口径。

边界 3（§A1 ActivationSource 语义）：枚举注释明确该字段反映"实际激活来源"
而非证据有无；新增测试覆盖"有 credible 证据但评分全不达标→走保底"混合路径。

边界 4（§A2 废弃重载）：基础 Evaluate(EccmSourceInfo, bool, proposals) 重载
在生产代码中零调用，标记 [[deprecated]]；测试文件整文件 pragma 抑制警告。

边界 5（§12 类头注释对齐）：RadarModule/EosModule/EsrModule 三处
buildExternalOutput 注释补足跨模块差异说明（SAR 因产品为图像不暴露此能力）。

文档：§6 标题从"原示例误导已修"修正为"底层非对称未变"，与正文诚实标注一致；
§1/§A1/§A2/§12 复核状态补充"边界清理"段落。

验证：1q_unit_tests 109 测试通过；public_api_boundary_guard/install_manifest
ctest 通过；ar_batch_validation 52 场景 warning=0 error=0 回放失败 0。

+++ include/1q/sar/session/SarCycleInput.h
@@ -99,7 +99,7 @@ struct ONEQ_API SarRawIqFrame {
  */
 struct ONEQ_API SarCycleInput {
   std::uint32_t cycle_index{0U};
-  double dt_sec{1.0};
+  float dt_sec{1.0f};
   SarPlatformState platform{};
   SarPointTargetList point_targets{};
   SarRawIqFrame raw_iq{};
