+++ include/1q/sar/session/SarCycleInputAdapter.h
@@ -52,7 +52,7 @@ struct ONEQ_API SarCycleInputAdapter {
   static bool Build(const SarPlatformState& platform,
                     const SarPointTargetList& targets,
                     const config::SarMissionConfig& mission,
-                    double dt_sec,
+                    float dt_sec,
                     const std::vector<SarExternalPulseInput>& external_pulses,
                     SarCycleInput* output,
                     SarCoordinateStatus* status = nullptr);
